requirements.txt
Flask==3.1.3
Flask-Cors == 6.0.2
tensorflow == 2.21.0
numpy == 2.4.3
Pillow == 12.1.1
gunicorn == 25.3.0
requests == 2.33.0
tensorflow == 2.21.0
